backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
openai
googleapi
dotenv